+++ Runtime/Components/AutoDestroy.cs
@@ -3,6 +3,7 @@ using UnityEngine;
 namespace NekoLib.Components
 {
     [DisallowMultipleComponent]
+    [AddComponentMenu("NekoLib/Auto Destroy")]
     public class AutoDestroy : MonoBehaviour
     {
         [SerializeField, Min(0f), Tooltip("Time in seconds before the object is destroyed")] private float _lifetime = 5f;